pyproject.toml
[project]
name = "agentic-ai-blog-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arxiv>=2.2.0",
    "chromadb>=1.0.15",
    "faiss-cpu>=1.11.0.post1",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-chroma>=0.2.5",
    "langchain-cli[inmem]>=0.0.36",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.6",
    "langchain-ollama>=0.3.6",
    "langchain-openai>=0.3.28",
    "langgraph>=0.6.0",
    "langgraph-cli[inmem]>=0.1.54",
    "python-dotenv>=1.1.1",
    "streamlit>=1.47.1",
    "tavily-python>=0.7.10",
    "uvicorn>=0.35.0",
    "watchdog>=6.0.0",
    "wikipedia>=1.4.0",
]
